Split app into separate files
